Revert coupon changes

+++ Cloudents.Core/DTOs/PaymentDetailDto.cs
@@ -7,7 +7,7 @@ namespace Cloudents.Core.DTOs
         public double TutorPricePerHour { get; set; }
         public string? CouponCode { get; set; }
         public CouponType? CouponType { get; set; }
-        public double? CouponValue { get; set; }
+        public decimal? CouponValue { get; set; }
 
         public long? CouponTutor { get; set; }
